Splitted tests for Empty and Default into separated files
